Add Entity and ticking
This commit adds a new class Entity along with variants of it, such as
PhysicsEntity. This is defined in common/entity.ts.

It also adds server ticking, which is handled by a level.Level. Ticking
will help the server calculate physics in a more trivial manner
(compared to tickless systems).

This comes with modifications to the types that are transferred over the
Websocket. More specifically, instead of a single "correction" event for
the player, the server can now correct multiple or all entities in a
map. The client is expected to lerp the positions to make everything
smooth while not having a physics engine of its own.

This commit also moves the anticheat code to a .bak folder to indicate
that it's not yet ready. It shall be revisited at a later time when the
physice engine's API is finalized.